edit request

+++ src/app/home-page/home-page.component.ts
@@ -15,8 +15,9 @@ import { HttpClient } from '@angular/common/http';
 export class HomePageComponent implements OnInit {
     filteredProduct: Product[] = [];
     products: Product[] = [];
-    notification;
     config: any;
+    create;
+    edit;
 
     constructor(
         private router: Router,
@@ -36,7 +37,6 @@ export class HomePageComponent implements OnInit {
         this.productService.getListProductsByUser().subscribe(
             next => {
                 this.filteredProduct = next;
-                // this.filteredProduct = this.filteredProduct;
             }
         );
         this.productService
@@ -48,25 +48,30 @@ export class HomePageComponent implements OnInit {
         this.config.currentPage = event;
     }
     search(key) {
-        this.notification = true;
         this.filteredProduct = this.filteredProduct.filter(product => product.name.toLowerCase().includes(key.toLowerCase()));
         console.log("list " + this.filteredProduct.length);
     }
 
     deletePost(i) {
-        for (let j = 0; j < this.filteredProduct.length - 1; j++) {
-            const product = this.filteredProduct[j];
-            if (product.id === i) {
-                console.log(product);
-                this.productService.deleteProduct(product.id).subscribe(() => {
-                    console.log("delete " + product.id);
-                    const indexOf = this.filteredProduct.indexOf(product);
-                    this.filteredProduct.splice(indexOf, 1);
-                    alert("Delete done");
-                    console.log("Delete");
-                });
+        var result = confirm("Bạn có chắc chắn xóa người dùng này?");
+        if (result == true) {
+            for (let j = 0; j < this.filteredProduct.length - 1; j++) {
+                const product = this.filteredProduct[j];
+                if (product.id === i) {
+                    console.log(product);
+                    this.productService.deleteProduct(product.id).subscribe(() => {
+                        console.log("delete " + product.id);
+                        const indexOf = this.filteredProduct.indexOf(product);
+                        this.filteredProduct.splice(indexOf, 1);
+                        alert("Delete done");
+                        console.log("Delete");
+                    });
+                }
             }
+        } else{
+            console.log("NO DELTE")
         }
+        
 
     }
 }
